requirements.txt
torch>=2.1
torchaudio>=2.1
transformers>=4.39
soundfile>=0.12
numpy>=1.24
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
